transform: Use sigs.k8s.io/yaml for marshal objects
- Use sigs.k8s.io/yaml to marshal go types into yaml with correct
values. kyaml marshal does not marshal some nested fields properly.
- Rename yaml to kyaml. yaml is sigs.k8s.io/yaml.
- Update DaemonSet ResourceRequirements transform to use yaml marshal
for simplicity. This also changes the behavior of ResourceRequirements
transform to not retain existing resources. This behavior is okay
because the input for resources is generally the whole object. Direct
assignment of the given resource is fine.
- Update resource transform tests to remove partial update tests due to
the change in behavior.