Create custom component config
Convert the project into a multigroup project and add new API for
operator config in group config.storageos.com. Since the cluster API
lacks a group, the v1 directory of the cluster API is at the root of
apis/ directory. The new operator config API is under
apis/config.storageos.com/v1/ because the project doesn't have a default
domain, which is used for grouping.
Add custom configurations in the new operator config API and use it in
the manager to load and use the custom configuration to configure the
controller manager and the controllers. Remove the default flags and
allow setting them via component config.